backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-google-genai
langchain-community
langsmith
pillow
python-dotenv
mlx-vlm
torch
